unify solution, rename projects per discussion
