+++ Client/Properties/AssemblyInfo.cs
@@ -1,0 +1,6 @@
+﻿using System.Reflection;
+
+[assembly: AssemblyTitle("Anti-Cheat Client")]
+[assembly: AssemblyDescription("")]
+[assembly: AssemblyVersion("7.3.0")]
+[assembly: AssemblyFileVersion("7.3.0")]
