+++ Client/assets/nuget-icon.png
