+++ Client/packages.config
@@ -1,0 +1,5 @@
+﻿<?xml version="1.0" encoding="utf-8"?>
+<packages>
+  <package id="Pathoschild.Stardew.ModBuildConfig" version="2.1.0" targetFramework="net452" />
+  <package id="Platonymous.PyTK" version="1.2.5-beta-20180618" targetFramework="net452" />
+</packages>
