+++ Server/Properties/AssemblyInfo.cs
@@ -1,0 +1,6 @@
+﻿using System.Reflection;
+
+[assembly: AssemblyTitle("Anti-Cheat Server")]
+[assembly: AssemblyDescription("")]
+[assembly: AssemblyVersion("7.2.0")]
+[assembly: AssemblyFileVersion("7.2.0")]
