--- ServerCode/ServerCode.sln
@@ -1,25 +1,0 @@
-﻿
-Microsoft Visual Studio Solution File, Format Version 12.00
-# Visual Studio 15
-VisualStudioVersion = 15.0.27703.2047
-MinimumVisualStudioVersion = 10.0.40219.1
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "ServerCode", "ServerCode\ServerCode.csproj", "{E9FB569C-5B16-4146-BD9D-46568548C6BF}"
-EndProject
-Global
-	GlobalSection(SolutionConfigurationPlatforms) = preSolution
-		Debug|Any CPU = Debug|Any CPU
-		Release|Any CPU = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{E9FB569C-5B16-4146-BD9D-46568548C6BF}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{E9FB569C-5B16-4146-BD9D-46568548C6BF}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{E9FB569C-5B16-4146-BD9D-46568548C6BF}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{E9FB569C-5B16-4146-BD9D-46568548C6BF}.Release|Any CPU.Build.0 = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(SolutionProperties) = preSolution
-		HideSolutionNode = FALSE
-	EndGlobalSection
-	GlobalSection(ExtensibilityGlobals) = postSolution
-		SolutionGuid = {F20D7B2A-AD3F-4003-B061-B670C6E5A66A}
-	EndGlobalSection
-EndGlobal
