--- ServerCode/ServerCode/ServerCode/manifest.json
@@ -1,10 +1,0 @@
-{
-  "Name": "ServerCode",
-  "Author": "funnysnek",
-  "Version": "7.3.0",
-  "Description": "code for the server",
-  "UniqueID": "funnysnek.servercode",
-  "EntryDll": "ServerCode.dll",
-  "MinimumApiVersion": "2.0",
-  "UpdateKeys": []
-}
