--- anticheatviachat/anticheatviachat.sln
@@ -1,25 +1,0 @@
-﻿
-Microsoft Visual Studio Solution File, Format Version 12.00
-# Visual Studio 15
-VisualStudioVersion = 15.0.27703.2047
-MinimumVisualStudioVersion = 10.0.40219.1
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "anticheatviachat", "anticheatviachat\anticheatviachat.csproj", "{00F22AE1-8E7B-4627-B04E-14C75BF03CF2}"
-EndProject
-Global
-	GlobalSection(SolutionConfigurationPlatforms) = preSolution
-		Debug|Any CPU = Debug|Any CPU
-		Release|Any CPU = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{00F22AE1-8E7B-4627-B04E-14C75BF03CF2}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{00F22AE1-8E7B-4627-B04E-14C75BF03CF2}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{00F22AE1-8E7B-4627-B04E-14C75BF03CF2}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{00F22AE1-8E7B-4627-B04E-14C75BF03CF2}.Release|Any CPU.Build.0 = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(SolutionProperties) = preSolution
-		HideSolutionNode = FALSE
-	EndGlobalSection
-	GlobalSection(ExtensibilityGlobals) = postSolution
-		SolutionGuid = {34259797-19D9-4DDA-976A-A8F32C36078A}
-	EndGlobalSection
-EndGlobal
