--- anticheatviachat/anticheatviachat/anticheatviachat/manifest.json
@@ -1,10 +1,0 @@
-{
-   "Name": "anticheatviachat",
-   "Author": "funnysnek",
-   "Version": "7.2.0",
-   "Description": "anticheatviachat",
-   "UniqueID": "anticheatviachat.anticheatviachat",
-   "EntryDll": "anticheatviachat.dll",
-   "MinimumApiVersion": "2.0",
-   "UpdateKeys": []
-}
